Ensure push constants don't exceed 128 bytes
This is the guaranteed minimum maxPushConstantsSize.
