+++ externals/ShaderProcessor/include/ShaderProcessor/ShaderSource.hpp
@@ -2,6 +2,7 @@
 
 #include <cstdint>
 #include <string>
+#include <cassert>
 
 /// Contains source code and reflection info for a shader.
 struct ShaderSource {
@@ -76,3 +77,31 @@ struct ShaderSource {
     /// Reflection info.
     ReflectionInfo reflectionInfo;
 };
+
+/// Get the size of a push constant type.
+/**
+ * @param type The type of the push constant.
+ *
+ * @return The size of the type in bytes.
+ */
+static uint32_t PushConstantTypeToSize(ShaderSource::ReflectionInfo::PushConstant::Type type) {
+    switch (type) {
+    case ShaderSource::ReflectionInfo::PushConstant::Type::INT:
+        return sizeof(int32_t);
+    case ShaderSource::ReflectionInfo::PushConstant::Type::UINT:
+        return sizeof(uint32_t);
+    case ShaderSource::ReflectionInfo::PushConstant::Type::FLOAT:
+        return sizeof(float);
+    case ShaderSource::ReflectionInfo::PushConstant::Type::VEC2:
+        return sizeof(float) * 2;
+    case ShaderSource::ReflectionInfo::PushConstant::Type::VEC3:
+        return sizeof(float) * 3;
+    case ShaderSource::ReflectionInfo::PushConstant::Type::VEC4:
+        return sizeof(float) * 4;
+    case ShaderSource::ReflectionInfo::PushConstant::Type::MAT4:
+        return sizeof(float) * 4 * 4;
+    default:
+        assert(false);
+        return 0;
+    }
+}
