+++ src/Video/LowLevelRenderer/Interface/ShaderProgram.hpp
@@ -25,35 +25,6 @@ class ShaderProgram {
     /// Destructor.
     virtual ~ShaderProgram() {}
 
-  protected:
-    /// Get the size of a push constant type.
-    /**
-     * @param type The type of the push constant.
-     *
-     * @return The size of the type in bytes.
-     */
-    static uint32_t PushConstantTypeToSize(ShaderSource::ReflectionInfo::PushConstant::Type type) {
-        switch (type) {
-        case ShaderSource::ReflectionInfo::PushConstant::Type::INT:
-            return sizeof(int32_t);
-        case ShaderSource::ReflectionInfo::PushConstant::Type::UINT:
-            return sizeof(uint32_t);
-        case ShaderSource::ReflectionInfo::PushConstant::Type::FLOAT:
-            return sizeof(float);
-        case ShaderSource::ReflectionInfo::PushConstant::Type::VEC2:
-            return sizeof(glm::vec2);
-        case ShaderSource::ReflectionInfo::PushConstant::Type::VEC3:
-            return sizeof(glm::vec3);
-        case ShaderSource::ReflectionInfo::PushConstant::Type::VEC4:
-            return sizeof(glm::vec4);
-        case ShaderSource::ReflectionInfo::PushConstant::Type::MAT4:
-            return sizeof(glm::mat4);
-        default:
-            assert(false);
-            return 0;
-        }
-    }
-
   private:
     ShaderProgram(const ShaderProgram& other) = delete;
 };
